requirements.txt
Flask>=2.0
pandas>=1.3
gunicorn>=20.0
psycopg2-binary>=2.9 
google-cloud-storage>=2.0 # Para o ingest_data.py
requests>=2.25 # Se for baixar de URLs simples, o SDK do GCS é melhor